make pagination change handler dynamic
